data: move structures into its own module
